libafs/dir: Verify directory pathnames
Provide a new routine, GetVerifiedBlob() which will ensure that the
pathname contained within a directory blob is correctly terminated
before returning it to the caller. For the purposes of this function,
correct termination is defined as having a terminating \0 character
within the same directory page as the blob itself.

Reviewed-on: http://gerrit.openafs.org/5617
Reviewed-by: Derrick Brashear <shadow@dementix.org>
Tested-by: Derrick Brashear <shadow@dementix.org>
(cherry picked from commit 988e51520429195955921bb53a71974a0cdff0a8)

Change-Id: Ifb17d22a79b0b05a7dc209a8686b05225a4aefb0
Reviewed-on: http://gerrit.openafs.org/5746
Reviewed-by: Derrick Brashear <shadow@dementix.org>
Tested-by: Derrick Brashear <shadow@dementix.org>